Make file upload functions a WDIO custom command and add related Convention variables
